Show identified template GUID + GUI adjustments
Now GUID of the identified template is shown in the status text control.
The plan is to next change "Enroll" button to "Unenroll" button whenever
the identified template GUID is shown (in order to implement unenrolling
feature). In addition, this commit adjusts GUI layout a bit to make the
distances between neighbouring controls more consistent (i.e. 5px wide).